requirements/doc.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
alabaster==0.7.16
    # via sphinx
amqp==5.3.1
    # via
    #   -r requirements/test.txt
    #   kombu
appdirs==1.4.4
    # via
    #   -r requirements/test.txt
    #   fs
asgiref==3.8.1
    # via
    #   -r requirements/test.txt
    #   django
babel==2.16.0
    # via sphinx
billiard==4.2.1
    # via
    #   -r requirements/test.txt
    #   celery
boto3==1.36.2
    # via
    #   -r requirements/test.txt
    #   fs-s3fs
botocore==1.36.2
    # via
    #   -r requirements/test.txt
    #   boto3
    #   s3transfer
celery==5.4.0
    # via -r requirements/test.txt
certifi==2024.12.14
    # via
    #   -r requirements/test.txt
    #   requests
charset-normalizer==3.4.1
    # via
    #   -r requirements/test.txt
    #   requests
click==8.1.8
    # via
    #   -r requirements/test.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via
    #   -r requirements/test.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/test.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/test.txt
    #   celery
coverage[toml]==7.6.10
    # via
    #   -r requirements/test.txt
    #   pytest-cov
django==4.2.18
    # via
    #   -c https://raw.githubusercontent.com/openedx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/test.txt
    #   openedx-django-pyfs
doc8==1.1.2
    # via -r requirements/doc.in
docutils==0.21.2
    # via
    #   doc8
    #   readme-renderer
    #   restructuredtext-lint
    #   sphinx
edx-sphinx-theme==3.1.0
    # via -r requirements/doc.in
fs==2.4.16
    # via
    #   -r requirements/test.txt
    #   fs-s3fs
    #   openedx-django-pyfs
    #   xblock
fs-s3fs==1.1.1
    # via
    #   -r requirements/test.txt
    #   openedx-django-pyfs
idna==3.10
    # via
    #   -r requirements/test.txt
    #   requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via
    #   -r requirements/test.txt
    #   pytest
jinja2==3.1.5
    # via sphinx
jmespath==1.0.1
    # via
    #   -r requirements/test.txt
    #   boto3
    #   botocore
kombu==5.4.2
    # via
    #   -r requirements/test.txt
    #   celery
lazy==1.6
    # via
    #   -r requirements/test.txt
    #   xblock
lxml==5.3.0
    # via
    #   -r requirements/test.txt
    #   xblock
mako==1.3.8
    # via
    #   -r requirements/test.txt
    #   xblock
    #   xblock-utils
markupsafe==3.0.2
    # via
    #   -r requirements/test.txt
    #   jinja2
    #   mako
    #   xblock
nh3==0.2.20
    # via readme-renderer
openedx-django-pyfs==3.7.0
    # via
    #   -r requirements/test.txt
    #   xblock
packaging==24.2
    # via
    #   -r requirements/test.txt
    #   pytest
    #   sphinx
pbr==6.1.0
    # via stevedore
pluggy==1.5.0
    # via
    #   -r requirements/test.txt
    #   pytest
prompt-toolkit==3.0.48
    # via
    #   -r requirements/test.txt
    #   click-repl
pygments==2.19.1
    # via
    #   doc8
    #   readme-renderer
    #   sphinx
pytest==8.3.4
    # via
    #   -r requirements/test.txt
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.txt
pytest-django==4.9.0
    # via -r requirements/test.txt
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/test.txt
    #   botocore
    #   celery
    #   xblock
pytz==2024.2
    # via
    #   -r requirements/test.txt
    #   xblock
pyyaml==6.0.2
    # via
    #   -r requirements/test.txt
    #   xblock
readme-renderer==44.0
    # via -r requirements/doc.in
requests==2.32.3
    # via
    #   -r requirements/test.txt
    #   sphinx
restructuredtext-lint==1.4.0
    # via doc8
s3transfer==0.11.1
    # via
    #   -r requirements/test.txt
    #   boto3
simplejson==3.19.3
    # via
    #   -r requirements/test.txt
    #   xblock
    #   xblock-utils
six==1.17.0
    # via
    #   -r requirements/test.txt
    #   edx-sphinx-theme
    #   fs
    #   fs-s3fs
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
sphinx==3.5.3
    # via
    #   -r requirements/doc.in
    #   edx-sphinx-theme
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
sqlparse==0.5.3
    # via
    #   -r requirements/test.txt
    #   django
stevedore==5.4.0
    # via doc8
tzdata==2024.2
    # via
    #   -r requirements/test.txt
    #   celery
    #   kombu
urllib3==2.2.3
    # via
    #   -c https://raw.githubusercontent.com/openedx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/test.txt
    #   botocore
    #   requests
vine==5.1.0
    # via
    #   -r requirements/test.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements/test.txt
    #   prompt-toolkit
web-fragments==2.2.0
    # via
    #   -r requirements/test.txt
    #   xblock
    #   xblock-utils
webob==1.8.9
    # via
    #   -r requirements/test.txt
    #   xblock
xblock[django]==5.1.0
    # via
    #   -r requirements/test.txt
    #   xblock-utils
xblock-utils==4.0.0
    # via -r requirements/test.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
